Move rails_init into main rails2_ruby2 rather than rails/init.rb.
This is to prevent railties tests from failing because it will be automatically loaded by rails as a plugin and there is a test that asserts loaded plugins. This test does not expect rails2_ruby2 to show up.

This may be reverted later, but for now this keeps railties test suite green.

Command run in the rails2 directory to verify:

   rbenv local 2.3.8 && pushd railties && bundle exec --gemfile ../Gemfile.2.3 rake test ; popd